requirements.txt
pycocotools
requests
ipykernel
matplotlib
tensorflow
tensorflow-object-detection-api 
numpy 
Pillow
scikit-image 
opencv-python-headless